+++ src/main/java/net/commuty/parking/model/AccessRightAttributeName.java
@@ -16,7 +16,29 @@ public enum AccessRightAttributeName {
      * A unique {@link java.util.UUID} for each access.
      */
     @JsonProperty("id")
-    ID("id");
+    ID("id"),
+
+    /**
+     * When an access right is granted for a specific parking spot, this value represents its unique identifier ({@link java.util.UUID}). It is null otherwise.
+     * An access right may have a parking spot, according to its reason.
+     * Access rights with these reasons never owns a parking spot: <code>none</code>, <code>weekendAccess</code>, <code>spotReleased</code>
+     * Access rights with these reasons may have a parking spot: <code>spotReceived</code>, <code>permanentAccess</code>, <code>passenger</code>
+     */
+    @JsonProperty("parkingSpotId")
+    PARKING_SPOT_ID("parkingSpotId"),
+
+    /**
+     * This name is used for management.
+     */
+    @JsonProperty("parkingSpotName")
+    PARKING_SPOT_NAME("parkingSpotName"),
+
+    /**
+     * This display name is used for end-user. It can be the <code>parkingSiteName</code>parkingSiteName itself or another name (e.g. a zone name, a level name, etc).
+     */
+    @JsonProperty("parkingSpotDisplayName")
+    PARKING_SPOT_DISPLAY_NAME("parkingSpotDisplayName")
+    ;
 
     private final String attributeName;