+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.0] 2021-07-29
+
+* Support for flags `parkingSpotId`, `parkingSpotName`, `parkingSpotDisplayName` of `includeAttributes`  on `/v2/access-rights`.
+* New endpoint to list all the parking spots of a given parking site.
+
 ## [2.2.2] 2021-03-10
 
 * Ignore empty properties in AccessRights